requirements.txt
torch==1.9.0
adapter-transformers==2.1.1
datasets==1.9.0
ipykernel==6.0.2
jupyterlab==3.0.16
tokenizers==0.10.3
sklearn
tensorboard